# mydocslco/swagger.yaml
openapi: 3.0.0

info:
  title: Learn express and swagger
  description: LCO - a course segment about learning documentation
  version: 1.0.0
  contact:
    email: yash@google.com
    URL: "http://lco.dev"

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for local host
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - https
          - http
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: Returns a greet from LCO.
      responses:
        "200": # status code
          description: All good, success
          content:
            application/json:
              schema:
                type: string
                example: "mystring"
                items:
                  type: string
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error

  /lcoobject:
    get:
      tags:
        - Object
      summary: Returns a unique course.
      responses:
        "200": # status code
          description: All good, success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
                items:
                  type: string
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error

  /courses:
    get:
      tags:
        - Array
      summary: Returns an array of courses.
      responses:
        "200": # status code
          description: All good, success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error

  /mycourse/{courseId}:
    get:
      tags:
        - String
      summary: Returns course based on request id
      parameters:
        - name: courseId
          in: path
          required: true
          default: 22
          schema:
            type: string
      responses:
        "200": # status code
          description: All good, success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error

  /addCourse:
    post:
      tags:
        - String
      summary: adds a course to enxisting courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        "200": # status code
          description: All good, success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error

  /coursequery:
    get:
      tags:
        - String
      summary: trying to learn query
      parameters:
        - name: location
          in: query
          schema:
            type: string
            enum: [delhi, london, ahmedabad]
          required: true
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]

      responses:
        "200": # status code
          description: All good, success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error
  
  /courseupload:
    post:
      tags:
        - Files
      summary: uploading course image
      requestBody: 
        content:
          multipart/form-data:
            schema: 
              type: object
              properties:
                sampleFile: 
                  type: string
                  format: binary
      responses:
        "200": # status code
          description: All good, success
          
        "400": # status code
          description: Bad Request

        "500": # status code
          description: Internal server error